image_transformer/requirements.txt
Flask~=1.1.2
Pillow~=8.2.0
flask-cors~=3.0.10